ml_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
transformers==4.37.2
pillow==10.2.0
httpx==0.26.0
python-multipart==0.0.6
numpy==1.26.3
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
